requirements.txt
absl-py==1.0.0
androguard==3.3.5
asn1crypto==1.5.1
astroid==2.4.2
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
cachetools==5.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
Flask==2.1.2
Flask-MySQLdb==1.0.1
flatbuffers==1.12
fonttools==4.29.1
future==0.18.2
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.3.0
isort==5.6.4
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.1
jupyter-core==4.10.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lazy-object-proxy==1.4.3
libclang==14.0.1
lxml==4.8.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mysqlclient==2.1.0
nest-asyncio==1.5.5
networkx==2.8.1
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
Pygments==2.12.0
pylint==2.6.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
pyzmq==23.0.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.0
scipy==1.8.1
six==1.16.0
sklearn==0.0
stack-data==0.2.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
toml==0.10.2
tornado==6.1
traitlets==5.2.1.post0
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
wincertstore==0.2
wrapt==1.12.1
WTForms==3.0.1
zipp==3.8.0
